move input/output to io module which contains some generic things needed for both
