+++ src/io/output/attribute.rs
