+++ src/io/output/attribute_plugin.rs
