+++ src/lib.rs
@@ -14,10 +14,9 @@ pub mod communication;
 pub(crate) mod config;
 pub(crate) mod domain;
 pub(crate) mod initial_conditions;
-pub(crate) mod input;
+pub(crate) mod io;
 pub(crate) mod mass;
 pub(crate) mod named;
-pub(crate) mod output;
 pub(crate) mod parameters;
 pub(crate) mod particle;
 pub(crate) mod physics;
