+++ src/physics/mod.rs
@@ -12,11 +12,11 @@ pub use self::gravity::plugin::GravityPlugin;
 use self::parameters::Parameters;
 pub use self::time::Time;
 use crate::domain::ExchangeDataPlugin;
-use crate::input::DatasetInputPlugin;
+use crate::io::input::DatasetInputPlugin;
+use crate::io::output::AttributeOutputPlugin;
+use crate::io::output::DatasetOutputPlugin;
 use crate::mass::Mass;
 use crate::named::Named;
-use crate::output::AttributeOutputPlugin;
-use crate::output::DatasetOutputPlugin;
 use crate::parameters::ParameterPlugin;
 use crate::plugin_utils::get_parameters;
 use crate::plugin_utils::panic_if_already_added;
